requirements.txt
websocket-client
scikit-learn
numpy
python-dotenv
